dump_db can run on v10 or v12

+++ bin/dump_db.sh
@@ -7,25 +7,67 @@ INSTALL_DIR='___INSTALL_DIR___'
 PRG=$( basename "$0" )
 YES=0
 NO=1
-AUTO_YES=${NO}
 
-dump_db() {
+
+get_ds_version() {
+  if [[ -f /sbin/db2ldif ]] ; then
+    DS_VERSION=10
+  elif [[ -f /usr/sbin/dsconf ]] ; then
+    DS_VERSION=12
+  else
+    die "Unable to determine version of ds389"
+  fi
+}
+
+
+db_dumper() {
   local _ts_start=$SECONDS
-  _dsctl db2ldif userRoot
+  "${@}"
   local _ts_end=$SECONDS
   local _elapsed=$( bc <<< "$_ts_end - $_ts_start" )
   echo "Database backup took: ${_elapsed} secs"
 
   local _ldif_out_fn=$( ls -t "${DS_LDIF_DIR}" | head -1)
+  [[ -f "${_ldif_out_fn}" ]] || die "unable to find db dump ldif"
+
   local _src="${DS_LDIF_DIR}"/"${_ldif_out_fn}"
-  local _tgt=/tmp/replcheck_"${HOST}"."${_ldif_out_fn}"
-  # ln -s -r "${DS_LDIF_DIR}"/"${_ldif_out_fn}" "${DS_LDIF_DIR}"/replcheck_${HOST}.ldif
+  local _tgt=/tmp/replcheck."${_ldif_out_fn}"
   mv "${_src}" "${_tgt}"
   chmod o+r "${_tgt}"
   echo  "Bkup LDIF: '${_tgt}'"
 
   echo -n 'Entries in DB: '
   grep '^dn: ' "${_tgt}" | wc -l
+}
+
+
+dump_db() {
+  if [[ "${DS_VERSION}" -eq 10 ]] ; then
+    cmd='/sbin/db2ldif'
+    parms=( '-s' "${DS_SUFFIX}" )
+  else
+    cmd='_dsctl'
+    parms=( db2ldif userRoot )
+  fi
+  db_dumper "${cmd}" "${parms[@]}"
+}
+
+
+stop_service() {
+  if [[ "${DS_VERSION}" -eq 10 ]] ; then
+    systemctl stop "${LDAP_SERVICE_NAME}"
+  else
+    _dsctl stop
+  fi
+}
+
+
+start_service() {
+  if [[ "${DS_VERSION}" -eq 10 ]] ; then
+    systemctl start "${LDAP_SERVICE_NAME}"
+  else
+    _dsctl start
+  fi
 }
 
 
@@ -37,10 +79,11 @@ purge_old() {
 print_usage() {
   echo
   cat <<ENDHERE
-${PRG} [OPTIONS] <ACTION>
+${PRG} [OPTIONS]
   OPTIONS
-    -h | --help   Print this help msg
-    -y | --yes     Answer Yes to all questions
+    -h | --help     Print this help msg
+    -y | --yes      Answer Yes to all questions (use in conjunction with --offline)
+    -o | --offline  Stop the server during the dump
 ENDHERE
   echo
 }
@@ -52,12 +95,20 @@ ENDHERE
 ###
 #
 # Process options
+AUTO_YES=${NO}
+OFFLINE=${NO}
 ENDWHILE=${NO}
 while [[ $# -gt 0 ]] && [[ $ENDWHILE -eq ${NO} ]] ; do
   case $1 in
-    -h|--help) print_usage; exit 0;;
+    -h|--help)
+      print_usage
+      exit 0
+      ;;
     -y | --yes)
       AUTO_YES=${YES}
+      ;;
+    -o | --offline)
+      OFFLINE=${YES}
       ;;
     --)
       ENDWHILE=${YES}
@@ -73,15 +124,13 @@ while [[ $# -gt 0 ]] && [[ $ENDWHILE -eq ${NO} ]] ; do
   shift
 done
 
-ACTION="${1}"
-
-[[ ${AUTO_YES} -ne ${YES} ]] \
+[[ ${OFFLINE} -eq ${YES} && ${AUTO_YES} -ne ${YES} ]] \
 && continue_or_exit "DS389 will be stopped during the backup. Continue?"
 
 purge_old
 
-_dsctl stop
+# [[ ${OFFLINE} -eq ${YES} ]] && stop_service
 
 dump_db
 
-_dsctl start
+# [[ ${OFFLINE} -eq ${YES} ]] && start_service
